refactor: move to a clean folder structure
