+++ src/components/Nav.tsx
@@ -1,7 +1,7 @@
 "use client";
 
-import { navlinks } from "@/contstants/navlinks";
-import { cn } from "@/utils/cn";
+import { navlinks } from "@/constants/navlinks";
+import { cn } from "@/lib/utils";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import React from "react";
