+++ src/components/sections/Home/ContainerLayout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/utils/cn";
+import { cn } from "@/lib/utils";
 import React from "react";
 import { Layout } from "./BoxComponents";
 import { MobileLayout } from "./MobileBoxComponents";
